Adjust coding structure.

+++ src/loaders.py
@@ -1,7 +1,7 @@
 import os, json, zipfile, io, requests, subprocess
 from .models import BaseClientModel
 from pathlib import Path
-from typing import Dict, Any, Tuple, Optional, List
+from typing import Dict, Any, Tuple, Optional, List, Union
 
 
 class ModLoaderInstaller(BaseClientModel):
@@ -19,6 +19,26 @@ class ModLoaderInstaller(BaseClientModel):
         self._minecraft_manifest = requests.request(
             "GET", "https://launchermeta.mojang.com/mc/game/version_manifest.json"
         ).json()
+
+    def _set_paths(
+        self, version_name: str, minecraft_dir_path: Optional[str] = None
+    ) -> None:
+        if minecraft_dir_path is None:
+            self.minecraft_dir_path = self._get_minecraft_dir_path()
+        self.libraries_path = os.path.join(minecraft_dir_path, "libraries")
+        self.versions_path = os.path.join(minecraft_dir_path, "versions")
+        self.profile_path = os.path.join(self.versions_path, version_name)
+        self.profile_json_file = os.path.join(self.profile_path, f"{version_name}.json")
+
+    def _write_version_file(self, data: Dict[str, Any]) -> bool:
+        try:
+            os.makedirs(self.profile_path, exist_ok=True)
+            with open(self.profile_json_file, "w", encoding="utf-8") as file:
+                json.dump(data, file, indent=2, ensure_ascii=False)
+        except Exception as e:
+            print(f"写入配置文件失败: {e}")
+            return False
+        return True
 
     def _get_minecraft_dir_path() -> str:
         home_path = Path.home()
@@ -44,6 +64,30 @@ class ModLoaderInstaller(BaseClientModel):
             minecraft_dir_path = os.path.join(str(home_path), ".minecraft")
         return minecraft_dir_path
 
+    def _resolve_maven_coord(self, coord: str) -> str:
+        parts = coord.replace("[", "").replace("]", "").split(":")
+        if len(parts) < 3:
+            raise ValueError(f"Invalid maven coord: {coord}")
+        group = parts[0]
+        artifact = parts[1]
+        version = parts[2]
+        classifier = parts[3] if len(parts) > 3 else None
+        extension = "jar"
+
+        if classifier and "@" in classifier:
+            classifier, extension = classifier.split("@")
+
+        group_path = group.replace(".", "/")
+        filename = f"{artifact}-{version}"
+        if classifier:
+            filename += f"-{classifier}"
+        filename += f".{extension}"
+
+        return f"{group_path}/{artifact}/{version}/{filename}"
+
+    def get_installer(self, url: str) -> Union[bytes, str]:
+        pass
+
     def install(
         self,
         minecraft_version: str,
@@ -63,58 +107,17 @@ class FabricInstaller(ModLoaderInstaller):
     ) -> None:
         super(FabricInstaller, self).__init__(api_base_url, max_workers)
 
-    def _maven_name_to_path(self, name: str) -> str:
-        parts = name.split(":")
-        if len(parts) != 3:
-            raise ValueError(f"Invalid Maven coordinate format: {name}")
-        group_id, artifact_id, version = parts
-        group_path = group_id.replace(".", "/")
-        return f"{group_path}/{artifact_id}/{version}/{artifact_id}-{version}.jar"
-
     def _get_lib_hash(self, data: Dict[str, Any]) -> Optional[Tuple[str, str]]:
-        if data.get("md5") is not None:
-            return data.get("md5"), "md5"
-        elif data.get("sha1") is not None:
+        if data.get("sha1") is not None:
             return data.get("sha1"), "sha1"
+        elif data.get("md5") is not None:
+            return data.get("md5"), "md5"
         elif data.get("sha256") is not None:
             return data.get("sha256"), "sha256"
         elif data.get("sha512") is not None:
             return data.get("sha512"), "sha512"
         else:
             return None
-
-    def _download_dependencies(
-        self,
-        minecraft_version: str,
-        loader_version: str,
-        dest_path: Optional[str] = None,
-        block_size: int = 8192,
-    ) -> List[bool]:
-        if dest_path is None:
-            home_path = Path.home()
-            dest_path = os.path.join(home_path, "Downloads")
-        response = self.get(
-            f"/v2/versions/loader/{minecraft_version}/{loader_version}/profile/json"
-        ).json()
-        libraries = response["libraries"]
-        grouped_tasks: Dict[str, List[str]] = {}
-        for lib in libraries:
-            hash_pack = self._get_lib_hash(lib)
-            maven_path = self._maven_name_to_path(lib.get("name"))
-            url = lib.get("url") + maven_path
-            folder_prefix = str(Path(maven_path).parent)
-            file_name = Path(maven_path).name
-            if hash_pack is not None:
-                task = (file_name, url, hash_pack[0], hash_pack[1])
-            else:
-                task = (file_name, url)
-            grouped_tasks.setdefault(folder_prefix, []).append(task)
-
-        results = []
-        for prefix, tasks in grouped_tasks.items():
-            path = os.path.join(dest_path, prefix)
-            results.extend(self.batch_download(tasks, path, block_size))
-        return results
 
     def install(
         self,
@@ -126,36 +129,42 @@ class FabricInstaller(ModLoaderInstaller):
     ):
         if version_name is None:
             version_name = f"fabric-{minecraft_version}-{loader_version}"
-        if minecraft_dir_path is None:
-            minecraft_dir_path = self._get_minecraft_dir_path()
-        libraries_path = os.path.join(minecraft_dir_path, "libraries")
-        versions_path = os.path.join(minecraft_dir_path, "versions")
-        profile_path = os.path.join(versions_path, version_name)
-        profile_json_file = os.path.join(profile_path, f"{version_name}.json")
-        deps_res = self._download_dependencies(
-            minecraft_version, loader_version, libraries_path, download_block_size
-        )
-        if all(deps_res) is not True:
-            return False
-        os.makedirs(profile_path, exist_ok=True)
-        response = self.get(
+        self._set_paths(version_name, minecraft_dir_path)
+
+        version_profile = self.get(
             f"/v2/versions/loader/{minecraft_version}/{loader_version}/profile/json"
         ).json()
-        response["id"] = version_name
-        try:
-            with open(profile_json_file, "w", encoding="utf-8") as file:
-                json.dump(response, file, indent=2, ensure_ascii=False)
-        except Exception as e:
-            print(f"写入配置文件失败: {e}")
+        libraries = version_profile["libraries"]
+
+        grouped_tasks: Dict[str, List[str]] = {}
+        for lib in libraries:
+            hash_pack = self._get_lib_hash(lib)
+            maven_path = self._resolve_maven_coord(lib.get("name"))
+            url = lib.get("url") + maven_path
+            folder_prefix = str(Path(maven_path).parent)
+            file_name = Path(maven_path).name
+            if hash_pack is not None:
+                task = (file_name, url, hash_pack[0], hash_pack[1])
+            else:
+                task = (file_name, url)
+            grouped_tasks.setdefault(folder_prefix, []).append(task)
+
+        deps_res = []
+        for prefix, tasks in grouped_tasks.items():
+            path = os.path.join(self.libraries_path, prefix)
+            deps_res.extend(self.batch_download(tasks, path, download_block_size))
+        if all(deps_res) is not True:
             return False
-        return True
+
+        version_profile["id"] = version_name
+        return self._write_version_file(version_profile)
 
 
 class NeoForgeInstaller(ModLoaderInstaller):
     def __init__(
-        self, api_base_url: str = "https://maven.neoforged.net", max_workers=5
+        self, maven_base_url: str = "https://maven.neoforged.net", max_workers=5
     ) -> None:
-        super(NeoForgeInstaller, self).__init__(api_base_url, max_workers)
+        super(NeoForgeInstaller, self).__init__(maven_base_url, max_workers)
 
     def _generate_version_file(
         self,
@@ -186,27 +195,6 @@ class NeoForgeInstaller(ModLoaderInstaller):
         version_data["inheritsFrom"] = minecraft_version
         return (version_data, profile_data["libraries"], profile_data["data"], bin_path)
 
-    def _resolve_maven_coord(self, coord: str) -> str:
-        parts = coord.replace("[", "").replace("]", "").split(":")
-        if len(parts) < 3:
-            raise ValueError(f"Invalid maven coord: {coord}")
-        group = parts[0]
-        artifact = parts[1]
-        version = parts[2]
-        classifier = parts[3] if len(parts) > 3 else None
-        extension = "jar"
-
-        if classifier and "@" in classifier:
-            classifier, extension = classifier.split("@")
-
-        group_path = group.replace(".", "/")
-        filename = f"{artifact}-{version}"
-        if classifier:
-            filename += f"-{classifier}"
-        filename += f".{extension}"
-
-        return f"{group_path}/{artifact}/{version}/{filename}"
-
     def _process(
         self,
         installer_tool_path: str,
@@ -217,7 +205,7 @@ class NeoForgeInstaller(ModLoaderInstaller):
         neoform_mapping_path: str,
         bin_path: bytes,
     ) -> bool:
-        if os.path.exists(patched_path) is not None:
+        if os.path.exists(patched_path) is not False:
             os.remove(patched_path)
         with open(".temp.lzma", "wb") as file:
             file.write(bin_path)
@@ -242,12 +230,23 @@ class NeoForgeInstaller(ModLoaderInstaller):
         ]
         try:
             subprocess.run(
-                command, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE
+                command,
+                check=True,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                text=True,
             )
+        except subprocess.CalledProcessError as e:
+            print(f"Fail to patch {minecraft_jar_path}: {e}")
+            print(f"Stdout: {e.stdout}")
+            print(f"Stderr: {e.stderr}")
+            return False
         except Exception as e:
             print(f"Fail to patch {minecraft_jar_path}: {e}")
             return False
-        os.remove(".temp.lzma")
+        finally:
+            if os.path.exists(".temp.lzma") is not False:
+                os.remove(".temp.lzma")
         return True
 
     def install(
@@ -260,18 +259,13 @@ class NeoForgeInstaller(ModLoaderInstaller):
     ) -> bool:
         if version_name is None:
             version_name = f"neoforge-{loader_version}"
-        if minecraft_dir_path is None:
-            minecraft_dir_path = self._get_minecraft_dir_path()
-        libraries_path = os.path.join(minecraft_dir_path, "libraries")
-        versions_path = os.path.join(minecraft_dir_path, "versions")
-        profile_path = os.path.join(versions_path, version_name)
-        profile_json_file = os.path.join(profile_path, f"{version_name}.json")
+        self._set_paths(version_name, minecraft_dir_path)
 
         version_profile, libraries, satic_data, bin_patch = self._generate_version_file(
             minecraft_version, loader_version, version_name
         )
-        grouped_tasks: Dict[str, List[str]] = {}
 
+        grouped_tasks: Dict[str, List[str]] = {}
         for lib in libraries:
             sha1_value = lib["downloads"]["artifact"]["sha1"]
             folder_prefix = str(Path(lib["downloads"]["artifact"]["path"]).parent)
@@ -298,18 +292,18 @@ class NeoForgeInstaller(ModLoaderInstaller):
 
         deps_res = []
         for prefix, tasks in grouped_tasks.items():
-            path = os.path.join(libraries_path, prefix)
+            path = os.path.join(self.libraries_path, prefix)
             deps_res.extend(self.batch_download(tasks, path, download_block_size))
         if not all(deps_res):
             return False
 
         if not os.path.exists(
-            Path(versions_path, minecraft_version, f"{minecraft_version}.jar")
+            Path(self.versions_path, minecraft_version, f"{minecraft_version}.jar")
         ):
             self.single_download(
                 version_data["downloads"]["client"]["url"],
                 f"{minecraft_version}.jar",
-                Path(versions_path, minecraft_version),
+                Path(self.versions_path, minecraft_version),
                 download_block_size,
                 version_data["downloads"]["client"]["sha1"],
                 "sha1",
@@ -323,32 +317,25 @@ class NeoForgeInstaller(ModLoaderInstaller):
 
         pro_res = self._process(
             Path(
-                libraries_path,
+                self.libraries_path,
                 self._resolve_maven_coord(install_tool_coord),
             ),
-            Path(versions_path, minecraft_version, f"{minecraft_version}.jar"),
-            Path(libraries_path, maven_path),
+            Path(self.versions_path, minecraft_version, f"{minecraft_version}.jar"),
+            Path(self.libraries_path, maven_path),
             Path(
-                libraries_path,
+                self.libraries_path,
                 self._resolve_maven_coord(satic_data["PATCHED"]["client"]),
             ),
-            libraries_path,
+            self.libraries_path,
             Path(
-                libraries_path,
+                self.libraries_path,
                 self._resolve_maven_coord(
                     f"net.neoforged:neoform:{satic_data["MCP_VERSION"]["client"]}:mappings@tsrg.lzma"
-                ),
+                ).replace("'", ""),
             ),
             bin_patch,
         )
         if pro_res is not True:
             return False
 
-        try:
-            os.makedirs(profile_path, exist_ok=True)
-            with open(profile_json_file, "w", encoding="utf-8") as file:
-                json.dump(version_profile, file, indent=2, ensure_ascii=False)
-        except Exception as e:
-            print(f"写入配置文件失败: {e}")
-            return False
-        return True
+        return self._write_version_file(version_profile)
